+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 # RxDB Changelog
 
 <!-- CHANGELOG NEWEST -->
+- FIX image attachments not working correctly in the browser
 - FIX Push batchSize in not respected [#3994](https://github.com/pubkey/rxdb/issues/3994)
 - FIX boolean indexes not working [#3994](https://github.com/pubkey/rxdb/issues/3994)
 <!-- ADD new changes here! -->